requirements.txt
flask
openai
python-dotenv
chromadb
pypdf2